backend/requirements.txt
Flask==1.1.2
flask-cors==3.0.10
psycopg2-binary==2.8.6
